Deep validation added, myPlants page added with validation style edit

+++ app/Http/Controllers/PlantController.php
@@ -7,17 +7,44 @@ use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Route;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use mysql_xdevapi\Result;
 use phpDocumentor\Reflection\Location;
+
 
 class PlantController extends Controller
 {
+    public function myPlants()
+    {
+        $validation = Plant::with('user')
+            ->where('user_id', Auth::id())
+            ->where('active', 1)
+            ->count();
 
+
+        if ($validation > 3) {
+            $query = Plant::with('user')->where('user_id', Auth::id());
+            $plants = $query->get();
+
+
+            return view('myPlants',
+                compact('plants'));
+        } else {
+            return redirect()->route('plant.index');
+        }
+
+    }
 
     //
     public function index(Request $request)
     {
         $query = Plant::query();
         $categories = Category::all();
+
+        $validation = Plant::with('user')
+            ->where('user_id', Auth::id())
+            ->where('active', 1)
+            ->count();
 
         if ($request->filled('category')) {
             $query->whereHas('category', function ($q) use ($request) {
@@ -32,7 +59,7 @@ class PlantController extends Controller
 
         $plants = $query->get();
 
-        return view('plants', compact('plants', 'categories'));
+        return view('plants', compact('plants', 'categories', 'validation'));
     }
 
     public function show(Plant $plant)
@@ -51,9 +78,15 @@ class PlantController extends Controller
 
     }
 
-    public function edit(Plant $plant)
+    public function edit(string $id)
     {
         $categories = Category::all();
+        $plant = Plant::find($id);
+
+        if (Auth::user()->cannot('edit-check', $plant)) {
+            return redirect()->route('plant.index');
+        }
+
         return view('edit',
             compact('categories', 'plant'));
 
@@ -61,24 +94,30 @@ class PlantController extends Controller
 
     public function destroy(Plant $plant)
     {
+        if (Auth::user()->cannot('delete-check', $plant)) {
+            return redirect()->route('plant.index');
+        }
+
         $plant->delete();
 
         return redirect()
             ->route('plant.index');
+
     }
 
     public function store(Request $request)
     {
         $request->validate([
             'name' => 'required|max:100',
-            'description' => 'required'
+            'description' => 'required',
+            'category' => 'required'
         ]);
 
         $plant = new Plant();
         $plant->name = $request->input('name');
         $plant->description = $request->input('description');
         $plant->user_id = Auth::id();
-        $plant->category_id = $request->input('category_id');
+        $plant->category_id = $request->input('category');
         $plant->save();
 
         return redirect()->route('plant.index');
@@ -86,16 +125,10 @@ class PlantController extends Controller
 
     public function update(Request $request, Plant $plant)
     {
-
-
-        if (Auth::id() !== $plant->user_id) {
-            abort(route('plant.index'));
-        }
         $request->validate([
             'name' => 'required|max:100',
             'description' => 'required'
         ]);
-
 
         $plant->name = $request->input('name');
         $plant->description = $request->input('description');
